fix rubocop
