+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
           - `metastate` should set as a keyword argument
           - `long_press_keycode` as well
        - Before: `@driver.press_keycode 66, 1` (Can set only metastate)
+- [Internal] Change directory and file structure
 
 ### Bug fixes
