+++ lib/appium_lib_core/ios/xcuitest/device/performance.rb
@@ -21,7 +21,7 @@ module Appium
                 def get_performance_record(save_file_path: './performance', profile_name: 'Activity Monitor',
                                            remote_path: nil, user: nil, pass: nil, method: 'PUT')
                   option = ::Appium::Core::Device::ScreenRecord.new(
-                      remote_path: remote_path, user: user, pass: pass, method: method
+                    remote_path: remote_path, user: user, pass: pass, method: method
                   ).upload_option
 
                   option[:profileName] = profile_name
